Remove code from header file factory.h
